Changes for bambino.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@1481 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/InhibConn.cpp
@@ -18,7 +18,7 @@ InhibConn::InhibConn(const char * name, HyPerCol * hc, HyPerLayer * pre, HyPerLa
    this->name = strdup(name);
    this->parent = hc;
    this->nfPre = pre->clayer->numFeatures;
-   this->numBundles = this->nfPre;
+   this->numBundles = (int) this->nfPre;
 
    initialize(NULL, pre, post, CHANNEL_INH);
